style(driver): silence clippy in the grad_cli finite-difference helper
Annotate `finite_diff` (the central-difference reference driver) with
allow(too_many_arguments) and drop a redundant `as f64` on an already-f64
subtraction. No behavior change; the 9 grad_cli gates stay green.

+++ crates/mercury_driver/src/lib.rs
@@ -879,6 +879,7 @@ mod grad_cli_tests {
     }
 
     /// Central finite-difference gradient of `out[loss_out][0]` w.r.t. each `wrt` element.
+    #[allow(clippy::too_many_arguments)]
     fn finite_diff(
         prog: &Program,
         fwd_name: Symbol,
@@ -903,7 +904,7 @@ mod grad_cli_tests {
                 let lp = loss(&mut bufs);
                 bufs[wi][j] = orig - eps;
                 let lm = loss(&mut bufs);
-                g[j] = (lp - lm) as f64 / (2.0 * eps as f64);
+                g[j] = (lp - lm) / (2.0 * eps as f64);
             }
             out.push(g);
         }
